requirements.txt
fire
pyyaml
h5py
Pillow
onnx
efficientnet_pytorch
onnxruntime
